+++ src/Suggest/Phrase.php
@@ -3,6 +3,7 @@
 namespace Elastica\Suggest;
 
 use Elastica\Suggest\CandidateGenerator\AbstractCandidateGenerator;
+use Elastica\Suggest\CandidateGenerator\DirectGenerator;
 
 /**
  * Class Phrase.
@@ -138,9 +139,21 @@ class Phrase extends AbstractSuggest
     /**
      * @return $this
      */
+    public function addDirectGenerator(DirectGenerator $generator): self
+    {
+        return $this->addParam('candidate_generator', $generator);
+    }
+
+    /**
+     * @deprecated since version 7.2.0, use the "addDirectGenerator()" method instead.
+     *
+     * @return $this
+     */
     public function addCandidateGenerator(AbstractCandidateGenerator $generator): Phrase
     {
-        return $this->setParam('candidate_generator', $generator);
+        \trigger_deprecation('ruflin/elastica', '7.2.0', 'The "%s()" method is deprecated, use the "addDirectGenerator()" method instead. It will be removed in 8.0.', __METHOD__);
+
+        return $this->addParam('candidate_generator', $generator);
     }
 
     /**
@@ -153,13 +166,15 @@ class Phrase extends AbstractSuggest
         $baseName = $this->_getBaseName();
 
         if (isset($array[$baseName]['candidate_generator'])) {
-            $generator = $array[$baseName]['candidate_generator'];
+            $generators = $array[$baseName]['candidate_generator'];
             unset($array[$baseName]['candidate_generator']);
 
-            $keys = \array_keys($generator);
-            $values = \array_values($generator);
+            foreach ($generators as $generator) {
+                $keys = \array_keys($generator);
+                $values = \array_values($generator);
 
-            $array[$baseName][$keys[0]][] = $values[0];
+                $array[$baseName][$keys[0]][] = $values[0];
+            }
         }
 
         return $array;