+++ tests/Suggest/PhraseTest.php
@@ -8,12 +8,15 @@ use Elastica\Suggest;
 use Elastica\Suggest\CandidateGenerator\DirectGenerator;
 use Elastica\Suggest\Phrase;
 use Elastica\Test\Base as BaseTest;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 
 /**
  * @internal
  */
 class PhraseTest extends BaseTest
 {
+    use ExpectDeprecationTrait;
+
     /**
      * @group unit
      */
@@ -51,6 +54,33 @@ class PhraseTest extends BaseTest
         $phraseSuggest = new Phrase('suggest1', 'text');
         $phraseSuggest->setText('elasticsearch is bansai coor');
         $phraseSuggest->setAnalyzer('simple')->setHighlight('<suggest>', '</suggest>')->setStupidBackoffSmoothing(0.4);
+        $phraseSuggest->addDirectGenerator(new DirectGenerator('text'));
+        $suggest->addSuggestion($phraseSuggest);
+
+        $index = $this->_getIndexForTest();
+        $result = $index->search($suggest);
+        $suggests = $result->getSuggests();
+
+        // 3 suggestions should be returned: One in which both misspellings are corrected, and two in which only one misspelling is corrected.
+        $this->assertCount(3, $suggests['suggest1'][0]['options']);
+
+        $this->assertEquals('elasticsearch is <suggest>bonsai cool</suggest>', $suggests['suggest1'][0]['options'][0]['highlighted']);
+        $this->assertEquals('elasticsearch is bonsai cool', $suggests['suggest1'][0]['options'][0]['text']);
+    }
+
+    /**
+     * @group functional
+     *
+     * @group legacy
+     */
+    public function testPhraseSuggestWithBackwardsCompatibility(): void
+    {
+        $suggest = new Suggest();
+        $phraseSuggest = new Phrase('suggest1', 'text');
+        $phraseSuggest->setText('elasticsearch is bansai coor');
+        $phraseSuggest->setAnalyzer('simple')->setHighlight('<suggest>', '</suggest>')->setStupidBackoffSmoothing(0.4);
+
+        $this->expectDeprecation('Since ruflin/elastica 7.2.0: The "Elastica\Suggest\Phrase::addCandidateGenerator()" method is deprecated, use the "addDirectGenerator()" method instead. It will be removed in 8.0.');
         $phraseSuggest->addCandidateGenerator(new DirectGenerator('text'));
         $suggest->addSuggestion($phraseSuggest);